Add ClientServicesProvider to devtools

+++ packages/devtools/devtools/src/App.tsx
@@ -7,6 +7,7 @@ import React, { useState } from 'react';
 import { Box, CssBaseline, ThemeProvider } from '@mui/material';
 
 import { Client, fromDefaults, fromIFrame } from '@dxos/client';
+import { ClientServicesProvider } from '@dxos/client-services';
 import { Config, Defaults, Dynamics } from '@dxos/config';
 import { useAsyncEffect } from '@dxos/react-async';
 import { ClientContext } from '@dxos/react-client';
@@ -25,6 +26,7 @@ const Telemetry = () => {
 
 export const App = () => {
   const [client, setClient] = useState<Client>();
+  const [servicesProvider, setServicesProvider] = useState<ClientServicesProvider>();
 
   const onConfigChange = async (remoteSource?: string) => {
     if (client && client?.config.values.runtime?.client?.remoteSource === remoteSource) {
@@ -42,11 +44,13 @@ export const App = () => {
     const config = new Config(remoteSourceConfig, await Dynamics(), Defaults());
 
     {
-      if (client) {
+      if (client && servicesProvider) {
         setClient(undefined);
         await client.destroy();
+        await servicesProvider.close();
       }
-      const newClient = new Client({ config, services: remoteSource ? fromIFrame(config) : fromDefaults(config) });
+      setServicesProvider(remoteSource ? fromIFrame(config) : fromDefaults(config));
+      const newClient = new Client({ config, services: servicesProvider });
       await newClient.initialize();
       setClient(newClient);
     }
@@ -65,7 +69,7 @@ export const App = () => {
       <ThemeProvider theme={theme}>
         <CssBaseline />
         <FullScreen sx={{ flexDirection: 'row' }}>
-          <ClientContext.Provider value={{ client }}>
+          <ClientContext.Provider value={{ client, services: servicesProvider?.services }}>
             <Telemetry />
             <Box sx={{ display: 'flex', flex: 1, overflow: 'hidden' }}>
               <PanelsContainer sections={sections} />
